requirements.txt
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
bidict==0.21.2
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses==0.8
dnspython==1.16.0
eventlet==0.30.1
Flask==1.0.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
mock==4.0.3
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-engineio==2.0.3
python-socketio==1.8.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
scipy==1.5.4
six==1.15.0
tensorboard==1.13.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
torch==1.7.1
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
